consumer_service/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.1
djangorestframework==3.15.2
dnspython==2.6.1
idna==3.7
mongoengine==0.28.2
PyJWT==2.9.0
pymongo==4.8.0
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
